Upgrade to the latest version of hibernate, 3.5.6
GForge Tracker #29762
Peer reviewed by Ngoc

+++ software/caintegrator2-war/src/gov/nih/nci/caintegrator2/data/CaIntegrator2Dao.java
@@ -410,7 +410,7 @@ public interface CaIntegrator2Dao {
      * @param imageSeries to retrieve number of images.
      * @return number of images for source.
      */
-    int retrieveNumberImages(Collection<ImageSeries> imageSeries);
+    Long retrieveNumberImages(Collection<ImageSeries> imageSeries);
     
     /**
      * For a given genomic source, retrieves all distinct platforms.
